backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --no-hashes --no-annotate
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.31.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6 ; sys_platform == 'win32'
distro==1.9.0
fastapi==0.122.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-core==1.2.5
langchain-openai==1.1.0
langchain-postgres==0.0.16
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.47
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pgvector==0.3.6
psycopg==3.2.13
psycopg-binary==3.2.13 ; implementation_name != 'pypy'
psycopg-pool==3.2.8
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pypdf==6.4.0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2 ; sys_platform == 'win32'
urllib3==2.5.0
uuid-utils==0.12.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
